fix build on arm
svn path=/trunk/KDE/kdelibs/; revision=891099

+++ widgets/toolbutton.cpp
@@ -285,7 +285,7 @@ void ToolButton::paint(QPainter *painter,
             QPainter bufferPainter(&buffer);
             bufferPainter.setCompositionMode(QPainter::CompositionMode_DestinationIn);
             QColor alphaColor(Qt::black);
-            alphaColor.setAlphaF(qMin(0.95, d->opacity));
+            alphaColor.setAlphaF(qMin(qreal(0.95), d->opacity));
             bufferPainter.fillRect(buffer.rect(), alphaColor);
             bufferPainter.end();
